requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
autopep8==1.7.0
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.12.1
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
packaging==21.3
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.40
toml==0.10.2
werkzeug==2.1.2
